requirements.txt
# Core dependencies
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.3
python-dotenv==1.0.0

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
pymysql==1.1.0  # MySQL
aiosqlite==0.19.0  # Async SQLite

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0